final_project_ros_bootcamp
it supposed to record waypoints that i sent them from RViz and it needs to send recorded waypoint to the vehicle but it doesnt work properly. it records the goals in a waypoint array which it listens from /amcl_pose topic. Normally we sent those waypoints on move_base but unfortunately, vehicle doesnt get them. I will try to fix it soon after my midterm week.
And here is the link for a video that while we check /move_base/goal etc topics to see sended goals.
https://drive.google.com/file/d/1O4230lY5btpTrBZ8vzwjNeWDNO3tJXj0/view?usp=sharing